backend/requirements.txt
django==2.2.19
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0